# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Racetag Backend API
  version: 0.1.0
servers:
  - url: http://localhost:8600
    description: Local Development Server
security:
  - ApiKeyAuth: []
paths:
  /events/tag/batch:
    post:
      summary: Ingest a batch of tag events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagEventBatchDTO'
      responses:
        '200':
          description: Batch processed summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchIngestResultDTO'
  /classification:
    get:
      summary: Current classification standings
      description: Returns a snapshot of standings at request time. For live updates, subscribe to the Server-Sent Events stream at /stream.
      responses:
        '200':
          description: Standings ordered by race rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationDTO'
  /race:
    get:
      summary: Race state
      responses:
        '200':
          description: Current race metadata and participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaceDTO'
  /stream:
    get:
      summary: Server-Sent Events stream for live updates
      description: |
        Emits newline-delimited `data:` frames. Known event types:
        - `lap` – a tag has crossed the timing line.
        - `standings` – full standings snapshot (enriched with bib/name when a rider is registered).
        - `unknown_tag` – an `arrive` was received for a tag with no registered rider.
      responses:
        '200':
          description: SSE stream of lap, standings, and unknown_tag events
          content:
            text/event-stream:
              schema:
                type: string

  /riders:
    post:
      summary: Register or update a rider for a tag
      description: Upsert semantics — re-POSTing with the same tag_id updates bib/name but preserves created_at.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderCreateDTO'
      responses:
        '201':
          description: Rider created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDTO'
    get:
      summary: List all registered riders
      responses:
        '200':
          description: All riders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RidersListDTO'

  /riders/recent-reads:
    get:
      summary: Recent unknown-tag reads
      description: Returns the last N arrive events for tags that have no registered rider, newest first.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Recent unknown-tag reads in reverse-chronological order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentReadsListDTO'

  /riders/{tag_id}:
    get:
      summary: Get a rider by tag id
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDTO'
        '404':
          description: No rider registered for this tag
    delete:
      summary: Delete a rider
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Rider deleted
        '404':
          description: No rider registered for this tag
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    TagEventDTO:
      type: object
      required: [source, reader_ip, timestamp, event_type, tag_id]
      properties:
        source:
          type: string
          example: sirit-510
        reader_ip:
          type: string
        reader_serial:
          type: string
        timestamp:
          type: string
          description: ISO8601 with milliseconds and Z
          example: 2025-10-16T11:53:32.298Z
        event_type:
          type: string
          enum: [arrive, depart]
        tag_id:
          type: string
          description: Uppercase hex without 0x
        session_id:
          type: integer
          nullable: true
        antenna:
          type: integer
          nullable: true
        rssi:
          type: integer
          nullable: true
        first:
          type: string
          nullable: true
        last:
          type: string
          nullable: true

    TagEventBatchDTO:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TagEventDTO'

    BatchIngestResultDTO:
      type: object
      required: [events_processed]
      properties:
        events_processed:
          type: integer
  
    ParticipantDTO:
      type: object
      required: [tag_id, laps, finished]
      properties:
        tag_id:
          type: string
        laps:
          type: integer
        last_pass_time:
          type: string
          nullable: true
        finish_time:
          type: string
          nullable: true
        finished:
          type: boolean
        total_time_ms:
          type: integer
          nullable: true
        gap_ms:
          type: integer
          nullable: true
          description: Time gap to the leader in milliseconds when on the same lap (0 for leader; null if lapped)
        laps_behind:
          type: integer
          nullable: true
          description: Number of laps behind the leader (null or 0 if on same lap)
        bib:
          type: string
          nullable: true
          description: Rider bib number (null if no rider registered for this tag)
        name:
          type: string
          nullable: true
          description: Rider name (null if no rider registered for this tag)

    RiderDTO:
      type: object
      required: [tag_id, bib, name, created_at]
      properties:
        tag_id:
          type: string
          description: RFID tag id (uppercase hex, primary key)
        bib:
          type: string
          description: Bib number (stored as string to preserve leading zeros)
        name:
          type: string
          description: Rider display name
        created_at:
          type: string
          format: date-time
          description: UTC timestamp when the rider was first registered

    RiderCreateDTO:
      type: object
      required: [tag_id, bib, name]
      properties:
        tag_id:
          type: string
        bib:
          type: string
        name:
          type: string

    RidersListDTO:
      type: object
      required: [count, items]
      properties:
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RiderDTO'

    RecentReadDTO:
      type: object
      required: [tag_id, timestamp]
      properties:
        tag_id:
          type: string
        timestamp:
          type: string
          description: ISO8601 timestamp of the arrive event
        antenna:
          type: integer
          nullable: true
        rssi:
          type: integer
          nullable: true

    RecentReadsListDTO:
      type: object
      required: [count, items]
      properties:
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecentReadDTO'

    ClassificationDTO:
      type: object
      required: [count, standings]
      properties:
        count:
          type: integer
        standings:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantDTO'

    RaceDTO:
      type: object
      required: [total_laps, start_time, participants]
      properties:
        total_laps:
          type: integer
        start_time:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantDTO'