# apps/backend/racetag-backend/models_api.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-10-18T17:01:42+00:00
#
# Hand-maintained in sync with openapi.yaml (datamodel-codegen not installed).
# When editing openapi.yaml, update the models below to match.

from __future__ import annotations

from datetime import datetime
from enum import Enum
from typing import List, Optional

from pydantic import BaseModel, Field


class EventType(Enum):
    arrive = 'arrive'
    depart = 'depart'


class TagEventDTO(BaseModel):
    source: str = Field(..., example='sirit-510')
    reader_ip: str
    reader_serial: str | None = None
    timestamp: str = Field(
        ...,
        description='ISO8601 with milliseconds and Z',
        example='2025-10-16T11:53:32.298Z',
    )
    event_type: EventType
    tag_id: str = Field(..., description='Uppercase hex without 0x')
    session_id: Optional[int] = None
    antenna: Optional[int] = None
    rssi: Optional[int] = None
    first: Optional[str] = None
    last: Optional[str] = None


class TagEventBatchDTO(BaseModel):
    events: List[TagEventDTO]


class BatchIngestResultDTO(BaseModel):
    events_processed: int


class ParticipantDTO(BaseModel):
    tag_id: str
    laps: int
    last_pass_time: Optional[str] = None
    finish_time: Optional[str] = None
    finished: bool
    total_time_ms: Optional[int] = None
    gap_ms: Optional[int] = Field(
        None,
        description='Time gap to the leader in milliseconds when on the same lap (0 for leader; null if lapped)',
    )
    laps_behind: Optional[int] = Field(
        None, description='Number of laps behind the leader (null or 0 if on same lap)'
    )
    # Rider fields — populated from RiderStore at standings time (W-010)
    bib: Optional[str] = Field(None, description='Rider bib number (null if no rider registered for this tag)')
    name: Optional[str] = Field(None, description='Rider name (null if no rider registered for this tag)')


class ClassificationDTO(BaseModel):
    count: int
    standings: List[ParticipantDTO]


class RaceDTO(BaseModel):
    total_laps: int
    start_time: str
    participants: List[ParticipantDTO]


# ---------------------------------------------------------------------------
# Rider DTOs (W-010) — hand-written to match openapi.yaml RiderDTO /
# RiderCreateDTO schemas.
# ---------------------------------------------------------------------------

class RiderDTO(BaseModel):
    """Rider as returned by the API (includes server-assigned created_at)."""

    tag_id: str = Field(..., description='RFID tag id (uppercase hex, primary key)')
    bib: str = Field(..., description='Bib number (stored as string to preserve leading zeros)')
    name: str = Field(..., description='Rider display name')
    created_at: datetime = Field(..., description='UTC timestamp when the rider was first registered')


class RiderCreateDTO(BaseModel):
    """Payload accepted by POST /riders."""

    tag_id: str = Field(..., description='RFID tag id (uppercase hex)')
    bib: str = Field(..., description='Bib number')
    name: str = Field(..., description='Rider display name')


class RidersListDTO(BaseModel):
    """Response for GET /riders."""

    count: int
    items: List[RiderDTO]


class RecentReadDTO(BaseModel):
    """Single entry in the recent-unknown-tags ring buffer."""

    tag_id: str
    timestamp: str
    antenna: Optional[int] = None
    rssi: Optional[int] = None


class RecentReadsListDTO(BaseModel):
    """Response for GET /riders/recent-reads."""

    count: int
    items: List[RecentReadDTO]